+++ src/plotter/Data.java
@@ -11,10 +11,8 @@ public class Data {
     
     private Coordinates[] data;
     
-    Data(float dataIN[][]){
-        
-        this.data = new Coordinates[dataIN.length];
-        
+    Data(double dataIN[][]){
+        this.data = new Coordinates[dataIN.length];    
         for(int i = 0; i < dataIN.length; i++){
             this.data[i] = new Coordinates(dataIN[i][0], dataIN[i][1]);
         }
